bug(#502): scan the full XPath/pattern attribute set, not just test/select
count, string-length, name, translate, and redundant-double-negation only
scanned //@test | //@select, so a count(x) > 0 in an xsl:key @use, a
name() = 'x' in a @group-by, or a not(not(@a)) in a match pattern went
unreported. They now scan the same 17-attribute set the axis and predicate
linters already used.

That set was duplicated verbatim in the axis and predicate linters; it is
now the single source src/attributes.js (ATTRIBUTES/SELECTOR), imported by
all seven. node-set (@select only) and redundant-boolean-call (whole @test)
keep their deliberately narrow scope.

Each of the five packs gains a match-pattern occurrence proving the wider
scan; full suite, 100% coverage, ESLint, and xcop stay green.
