+++ src/string-length-linter.js
@@ -6,6 +6,7 @@
 const {nodes} = require('./xpath')
 const {comparedToZero} = require('./comparisons')
 const {metaOf, suppressed, defect} = require('./checks')
+const {SELECTOR} = require('./attributes')
 const {logger} = require('./logger')
 
 /**
@@ -122,7 +123,7 @@ const lintByStringLength = function(corpus, suppressions = []) {
   const defects = []
   if (!suppressed(CHECK, suppressions)) {
     for (const {file, xsl} of corpus) {
-      for (const attribute of nodes(xsl, '//@test | //@select')) {
+      for (const attribute of nodes(xsl, SELECTOR)) {
         for (const {offset, value, replacement} of comparisons(
           attribute.nodeValue,
         )) {
