+++ test/resources/name-packs/name-compared-to-string.yaml
@@ -3,8 +3,8 @@
 ---
 pack: name-compared-to-string
 found:
-  amount: 4
-  positions: [ [ 4, 36 ], [ 5, 36 ], [ 6, 36 ], [ 7, 36 ] ]
+  amount: 5
+  positions: [ [ 4, 36 ], [ 5, 36 ], [ 6, 36 ], [ 7, 36 ], [ 9, 26 ] ]
 input: |-
   <?xml version="1.0"?>
   <xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform" version="2.0" id="s">
@@ -13,5 +13,8 @@ input: |-
       <xsl:variable name="b" select="name() != 'q'"/>
       <xsl:variable name="c" select="local-name() = 'r'"/>
       <xsl:variable name="d" select="'s' = name()"/>
+    </xsl:template>
+    <xsl:template match="*[name() = 't']">
+      <hit/>
     </xsl:template>
   </xsl:stylesheet>
