+++ test/resources/string-length-packs/string-length-compared-to-zero.yaml
@@ -3,8 +3,8 @@
 ---
 pack: string-length-compared-to-zero
 found:
-  amount: 6
-  positions: [ [ 4, 36 ], [ 5, 36 ], [ 6, 36 ], [ 7, 36 ], [ 8, 36 ], [ 9, 36 ] ]
+  amount: 7
+  positions: [ [ 4, 36 ], [ 5, 36 ], [ 6, 36 ], [ 7, 36 ], [ 8, 36 ], [ 9, 36 ], [ 11, 29 ] ]
 input: |-
   <?xml version="1.0"?>
   <xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform" version="2.0" id="s">
@@ -15,5 +15,8 @@ input: |-
       <xsl:variable name="d" select="string-length(@x) &lt;= 0"/>
       <xsl:variable name="e" select="string-length(@x) &gt;= 1"/>
       <xsl:variable name="f" select="string-length(@x) &lt; 1"/>
+    </xsl:template>
+    <xsl:template match="item[string-length(@id) &gt; 0]">
+      <hit/>
     </xsl:template>
   </xsl:stylesheet>
