+++ test/resources/translate-packs/translate-for-case.yaml
@@ -3,8 +3,8 @@
 ---
 pack: translate-for-case
 found:
-  amount: 3
-  positions: [ [ 4, 36 ], [ 5, 36 ], [ 6, 36 ] ]
+  amount: 4
+  positions: [ [ 4, 36 ], [ 5, 36 ], [ 6, 36 ], [ 8, 28 ] ]
 input: |-
   <?xml version="1.0"?>
   <xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform" version="2.0" id="s">
@@ -12,5 +12,8 @@ input: |-
       <xsl:variable name="a" select="translate(@x, 'ABCDEFGHIJKLMNOPQRSTUVWXYZ', 'abcdefghijklmnopqrstuvwxyz')"/>
       <xsl:variable name="b" select="translate(@y, 'abcdefghijklmnopqrstuvwxyz', 'ABCDEFGHIJKLMNOPQRSTUVWXYZ')"/>
       <xsl:variable name="c" select="translate(concat(@p, @q), 'ABCDEFGHIJKLMNOPQRSTUVWXYZ', 'abcdefghijklmnopqrstuvwxyz')"/>
+    </xsl:template>
+    <xsl:template match="row[translate(@a, 'ABCDEFGHIJKLMNOPQRSTUVWXYZ', 'abcdefghijklmnopqrstuvwxyz')]">
+      <hit/>
     </xsl:template>
   </xsl:stylesheet>
